+++ src/mapping/patch_writer.hh
@@ -23,7 +23,7 @@ public:
 	void setPatchDesc(std::string patchDesc);
 	void setCableList(std::vector<CableMap> &cables);
 	void setParamList(std::vector<ParamMap> &params);
-	void setMidiSettings(MIDI::ModuleIds &ids, MIDI::Settings const &settings);
+	void setMidiSettings(MIDI::Settings const &settings);
 	void setExpanders(ExpanderMappings const &exp);
 	void addModuleStateJson(rack::Module *module);
 
@@ -38,7 +38,7 @@ public:
 private:
 	void mapInputJack(CableMap &map);
 	void mapOutputJack(CableMap &map);
-	void mapMidiCVJack(CableMap &map);
+	void mapMidiCVJack(CableMap &map, uint32_t midi_chan);
 	void mapMidiGateJack(CableMap &map, unsigned midi_chan);
 	void mapMidiCCJack(CableMap &cable, unsigned midi_chan);
 
@@ -49,7 +49,6 @@ private:
 	PatchData pd;
 	int64_t hubModuleId = -1;
 
-	MIDI::ModuleIds midiModuleIds;
 	MIDI::Settings midiSettings;
 
 	ExpanderMappings expanders;